@@ -1,9 +1,9 @@
-import {BrowserModule}  from '@angular/platform-browser';
-import {NgModule}       from '@angular/core';
-import {FormsModule}    from '@angular/forms';
-import {HttpModule}     from '@angular/http';
-import {RouterModule}   from '@angular/router';
-import {Ng2Webstorage}  from 'ng2-webstorage';
+import {BrowserModule}        from '@angular/platform-browser';
+import {NgModule}             from '@angular/core';
+import {FormsModule}          from '@angular/forms';
+import {HttpClientModule}     from '@angular/common/http';
+import {RouterModule}         from '@angular/router';
+import {NgxWebstorageModule}  from 'ngx-webstorage';
 
 import {AppComponent}         from './component/app.component';
 import {LoginComponent}       from './component/login.component';
@@ -14,22 +14,27 @@ import {UsersComponent}       from './component/users.component';
 import {AppService}           from './service/app.service';
 import {XHRHandler}           from './service/xhrhandler.service';
 import {AppDataService}       from './service/appdata.service';
-import {ChatWebsocketService} from './service/chatwebsocket.service';
 
 @NgModule({
-  declarations: [AppComponent, LoginComponent, ChatComponent, ChatStreamComponent, UsersComponent],
+  declarations: [
+    AppComponent,
+    LoginComponent,
+    ChatComponent,
+    ChatStreamComponent,
+    UsersComponent
+  ],
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule, 
-    Ng2Webstorage,
+    HttpClientModule, 
+    NgxWebstorageModule.forRoot(),
     RouterModule.forRoot([
       {path: '', redirectTo: '/login', pathMatch: 'full'},
       {path: 'login', component: LoginComponent},
-      {path: 'home', component: ChatComponent}
+      {path: 'chat', component: ChatComponent}
     ])
   ],
-  providers: [AppService, XHRHandler, AppDataService, ChatWebsocketService],
+  providers: [AppService, XHRHandler, AppDataService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
